+++ gs/base/gxht.c
@@ -421,7 +421,7 @@ gx_dc_ht_binary_write(
     const gx_device_color *         pdevc,
     const gx_device_color_saved *   psdc0,
     const gx_device *               dev,
-    uint			    offset,
+    int64_t			    offset,
     byte *                          pdata,
     uint *                          psize )
 {
@@ -554,7 +554,7 @@ gx_dc_ht_binary_read(
     const gs_imager_state * pis,
     const gx_device_color * prior_devc,
     const gx_device *       dev,        /* ignored */
-    uint		    offset,
+    int64_t		    offset,
     const byte *            pdata,
     uint                    size,
     gs_memory_t *           mem )       /* ignored */
